consolidate PatternThumbnailPreview and ComponentsThumbnail to use ThumbnailPreview
