+++ polaris.shopify.com/src/components/ComponentThumbnail/ComponentThumbnail.module.scss
@@ -1,12 +1,3 @@
 .ComponentThumbnail {
   filter: brightness(97%);
-  border-radius: var(--border-radius-600);
-  border-bottom-right-radius: 0;
-  border-bottom-left-radius: 0;
-  overflow: hidden;
-}
-
-.Image {
-  width: 100%;
-  height: auto;
 }
