+++ polaris.shopify.com/src/components/GlobalSearch/GlobalSearch.tsx
@@ -13,7 +13,7 @@ import TokenList from '../TokenList';
 import {Dialog} from '@headlessui/react';
 import {KeyboardEventHandler} from 'react';
 import FoundationsThumbnail from '../FoundationsThumbnail';
-import PatternThumbnailPreview from '../PatternThumbnailPreview';
+import PatternThumbnailPreview from '../ThumbnailPreview';
 import ComponentThumbnail from '../ComponentThumbnail';
 const CATEGORY_NAMES: {[key in SearchResultCategory]: string} = {
   components: 'Components',
