--- polaris.shopify.com/src/components/PatternThumbnailPreview/PatternThumbnailPreview.module.scss
@@ -1,20 +1,0 @@
-.Preview {
-  aspect-ratio: 16/9;
-  border-radius: var(--border-radius-600);
-  border-bottom-right-radius: 0;
-  border-bottom-left-radius: 0;
-  overflow: hidden;
-  background: var(--p-surface-neutral-subdued);
-  padding: 0 9%;
-  position: relative;
-  &.NoInner {
-    padding: 0;
-  }
-  display: flex;
-  align-items: flex-end;
-  img {
-    width: 100%;
-    object-fit: cover;
-    height: 100%;
-  }
-}
