+++ polaris.shopify.com/src/components/ThumbnailPreview/ThumbnailPreview.module.scss
@@ -1,0 +1,7 @@
+.ThumbnailPreview {
+  aspect-ratio: 16/9;
+  border-radius: var(--border-radius-600);
+  overflow: hidden;
+  background: var(--p-surface-neutral-subdued);
+  position: relative;
+}
